pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vineland-runner"
version = "0.1.0"
description = "LLM agent evaluator using Vineland Adaptive Behavior Scales methodology"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "tenacity>=8.0",
  "jsonschema>=4.0",
  "tqdm>=4.66",
]

[project.scripts]
vineland-runner = "vineland_runner.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "respx>=0.20"]

[tool.setuptools.packages.find]
where = ["."]
include = ["vineland_runner*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
